refactor graphql fields; refactor tests
