+++ config/testConfig.js
@@ -22,6 +22,3 @@ global.assert = chai.assert
 
 // CSS modules setup
 require('src/server/configureCSSModules')()
-
-// use one database per-process b/c ava runs tests in parallel
-global.dbConfig = require('src/test/db').configureDB()
