+++ package.json
@@ -39,11 +39,12 @@
     "start": "npm run symlinks && npm run icons:fetch && node server",
     "symlinks": "./node_modules/.bin/babel-node ./scripts/createSymlinks",
     "symlinks:remove": "./node_modules/.bin/rimraf ./node_modules/src",
-    "test": "npm run symlinks && npm run lint && npm run test:run",
-    "test:ci": "npm run db:create && npm run test:cov",
-    "test:cov": "npm run test:cov:run && npm run test:cov:send",
-    "test:cov:run": "./node_modules/.bin/nyc --reporter=lcov npm run test:run",
-    "test:cov:send": "./node_modules/.bin/codeclimate-test-reporter < ./coverage/lcov.info",
+    "test": "npm run symlinks && npm run lint && npm run test:db && npm run test:run",
+    "test:ci": "NODE_ENV=test npm run db:create && NODE_ENV=test npm run db:migrate:up && npm run lint && npm run test:cov",
+    "test:cov": "NODE_ENV=test npm run test:cov:run && npm run test:cov:send",
+    "test:cov:run": "NODE_ENV=test ./node_modules/.bin/nyc --reporter=lcov npm run test:run",
+    "test:cov:send": "NODE_ENV=test ./node_modules/.bin/codeclimate-test-reporter < ./coverage/lcov.info",
+    "test:db": "NODE_ENV=test npm run db:drop && NODE_ENV=test npm run db:create && NODE_ENV=test npm run db:migrate:up",
     "test:run": "NODE_ENV=test ./node_modules/.bin/ava --serial",
     "workers": "npm run symlinks && ./node_modules/.bin/babel-node ./server/workers"
   },
