+++ server/graphql/mutations/__tests__/createInviteCode-test.js
@@ -1,24 +1,21 @@
 import test from 'ava'
 
 import {connect} from 'src/db'
-import {resetData, cleanupDB} from 'src/test/db'
 import {runMutation} from 'src/test/graphql'
 
 const r = connect()
 
-import api from '../mutation'
+import fields from '../index'
 
-const mutation = 'mutation($inviteCode: InputInviteCode!) { createInviteCode(inviteCode: $inviteCode) { id code active } }'
+const mutation = `
+  mutation($inviteCode: InputInviteCode!) {
+    createInviteCode(inviteCode: $inviteCode) {
+      id code active
+    }
+  }
+`
 
-test.before(async () => {
-  await resetData()
-})
-
-test.after(async () => {
-  await cleanupDB()
-})
-
-test.serial('createInviteCode: creates the invite code correctly', async t => {
+test.serial('creates the invite code correctly', async t => {
   t.plan(5)
 
   const inputInviteCode = {
@@ -29,12 +26,14 @@ test.serial('createInviteCode: creates the invite code correctly', async t => {
 
   await runMutation(
     mutation,
-    api,
+    fields,
     {inviteCode: inputInviteCode},
     {currentUser: {id: 'me', handle: 'me', roles: ['admin']}} // admin user
   )
 
-  const savedInviteCode = await r.table('inviteCodes').getAll(inputInviteCode.code, {index: 'code'}).nth(0)
+  const savedInviteCode = await r.table('inviteCodes')
+    .getAll(inputInviteCode.code, {index: 'code'})
+    .nth(0)
 
   t.is(savedInviteCode.code, inputInviteCode.code)
   t.is(savedInviteCode.description, inputInviteCode.description)
