+++ server/graphql/mutations/index.js
@@ -1,0 +1,13 @@
+import {default as createInviteCode} from './createInviteCode'
+import {default as deactivateUser} from './deactivateUser'
+import {default as reactivateUser} from './reactivateUser'
+import {default as updateUser} from './updateUser'
+import {default as updateUserAvatar} from './updateUserAvatar'
+
+export default {
+  createInviteCode,
+  deactivateUser,
+  reactivateUser,
+  updateUser,
+  updateUserAvatar,
+}
