+++ server/graphql/queries/index.js
@@ -1,0 +1,17 @@
+import {default as findUsers} from './findUsers'
+import {default as getActiveStatuses} from './getActiveStatuses'
+import {default as getInviteCodeByCode} from './getInviteCodeByCode'
+import {default as getUser} from './getUser'
+import {default as getUserById} from './getUserById'
+import {default as getUsersByHandles} from './getUsersByHandles'
+import {default as getUsersByIds} from './getUsersByIds'
+
+export default {
+  findUsers,
+  getActiveStatuses,
+  getInviteCodeByCode,
+  getUser,
+  getUserById,
+  getUsersByHandles,
+  getUsersByIds,
+}
