+++ server/graphql/schemas/index.js
@@ -1,0 +1,17 @@
+import {default as ActiveStatus} from './ActiveStatus'
+import {default as AuthProvider} from './AuthProvider'
+import {default as AuthProviders} from './AuthProviders'
+import {default as InputInviteCode} from './InputInviteCode'
+import {default as InputUser} from './InputUser'
+import {default as InviteCode} from './InviteCode'
+import {default as User} from './User'
+
+export default {
+  ActiveStatus,
+  AuthProvider,
+  AuthProviders,
+  InputInviteCode,
+  InputUser,
+  InviteCode,
+  User,
+}
