+++ server/graphql/types/index.js
@@ -1,0 +1,5 @@
+import {default as GraphQLPhoneNumber} from './GraphQLPhoneNumber'
+
+export default {
+  GraphQLPhoneNumber,
+}
